technographic-service/requirements.txt
fastapi
uvicorn
torch
transformers
accelerate
bitsandbytes
googlesearch-python
requests
